backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
ultralytics==8.0.43
python-dotenv==1.0.0
werkzeug==2.2.3 